requirements.txt
opencv-python
tensorflow
keras
requests
flask
Pillow
